meta: Removed weird u32, u16, s64, etc types.

+++ libraries/libfile/tar.cpp
@@ -47,7 +47,7 @@ uint tar_count(void *tarfile)
     {
         count++;
 
-        u32 size = get_file_size(header);
+        size_t size = get_file_size(header);
 
         header = (TARRawBlock *)((char *)header + ((size / 512) + 1) * 512);
 
@@ -67,7 +67,7 @@ bool tar_read(void *tarfile, TARBlock *block, uint index)
         if (header->name[0] == '\0')
             return false;
 
-        u32 size = get_file_size(header);
+        size_t size = get_file_size(header);
 
         header = (TARRawBlock *)((char *)header + ((size / 512) + 1) * 512);
